requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiolimiter==1.2.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
bce-python-sdk==0.9.23
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
dill==0.3.9
diskcache==5.6.3
Django==5.1.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.12
packaging==24.2
prompt_toolkit==3.0.48
propcache==0.2.1
pycryptodome==3.21.0
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
qianfan==0.4.12.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
starlette==0.41.3
tenacity==8.5.0
typer==0.14.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
faiss-cpu
websocket-client==1.6.4
sentence-transformers
python-dotenv==1.0.1
